refactor: share truncateHtml and apply truncation to direct messages too
truncateHtml was only applied to inline query results. Direct messages
hitting a long definition fell back to the "too legendary" placeholder
instead of truncating with a Read more link.

Move truncateHtml to formatter.ts so both paths share it. Apply the same
render-first / per-field budget strategy to buildDefinition in ud-bot.ts.
Move truncateHtml unit tests to formatter.test.ts.
