+++ src/features/definitions/formatter.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import formatter from "./formatter";
+import formatter, { truncateHtml } from "./formatter";
 
 describe("formatter", () => {
   it("bold wraps text in <b> tags", () => {
@@ -18,6 +18,46 @@ describe("formatter", () => {
 
   it("code wraps text in <code> tags", () => {
     expect(formatter.code("snippet")).toBe("<code>snippet</code>");
+  });
+
+  describe("truncateHtml", () => {
+    it("returns text unchanged when shorter than maxLength", () => {
+      expect(truncateHtml("hello world", 100, "...")).toBe("hello world");
+    });
+
+    it("returns text unchanged when exactly at maxLength", () => {
+      const text = "a".repeat(50);
+      expect(truncateHtml(text, 50, "...")).toBe(text);
+    });
+
+    it("truncates plain text and appends suffix", () => {
+      const result = truncateHtml("a ".repeat(100), 50, "...");
+      expect(result.length).toBeLessThanOrEqual(50);
+      expect(result.endsWith("...")).toBe(true);
+    });
+
+    it("backs up to the last word boundary", () => {
+      // target=6 → "aaa bb" → backs up to "aaa"
+      expect(truncateHtml("aaa bbb ccc", 9, "...")).toBe("aaa...");
+    });
+
+    it("removes an incomplete HTML tag at the truncation point", () => {
+      const text = 'some text <a href="https://example.com">link</a> more text here now';
+      const result = truncateHtml(text, 20, "...");
+      expect(result).not.toMatch(/<[^>]*$/);
+      expect(result.length).toBeLessThanOrEqual(20);
+    });
+
+    it("preserves a complete HTML tag that fits within the budget", () => {
+      const text = `before <a href="https://x.com">word</a> after that we have more words pushing past the limit`;
+      expect(truncateHtml(text, 50, "...")).not.toMatch(/<[^>]*$/);
+    });
+
+    it("handles text with no spaces gracefully", () => {
+      const result = truncateHtml("x".repeat(100), 20, "...");
+      expect(result.length).toBeLessThanOrEqual(20);
+      expect(result.endsWith("...")).toBe(true);
+    });
   });
 
   describe("compress / decompress", () => {
